+++ lib/runners/python.rb
@@ -1,9 +1,9 @@
 module Runners
   module Python
     def show_runtime_versions
-      capture3! "python", "--version"
-      capture3! "pip", "--version"
+      # NOTE: `python --version` fails if `.python-version` file exists.
       capture3! "pyenv", "--version"
+      capture3! "pyenv", "versions"
     end
   end
 end