+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 
 [Full diff](https://github.com/sider/runners/compare/0.13.0...HEAD)
 
+- [Flake8] Fix error if `.python-version` file exists [#602](https://github.com/sider/runners/pull/602)
+
 ## 0.13.0
 
 [Full diff](https://github.com/sider/runners/compare/0.12.0...0.13.0)